[Feat] #24 - Move temporary_restrict to chat package. (this feature needs chat_channel_id)
